+++ out_avail/toot.sh
@@ -57,15 +57,22 @@ function toot_send {
    
     # Get the image, if exists, then send the post
     if [ ! -z "${imgurl}" ];then
-        
-        Outfile=$(mktemp)
-        curl "${imgurl}" -o "${Outfile}" --max-time 60 --create-dirs -s
+        if [ -f "${imgurl}" ];then
+            Outfile="${imgurl}"
+        else
+            Outfile=$(mktemp)
+            curl "${imgurl}" -o "${Outfile}" --max-time 60 --create-dirs -s
+        fi
         if [ -f "${Outfile}" ];then
             loud "Image obtained, resizing."       
             if [ -f /usr/bin/convert ];then
                 /usr/bin/convert -resize 800x512\! "${Outfile}" "${Outfile}" 
             fi
-            Limgurl=$(echo "--media ${Outfile} --description \'An automated image pulled from the post\'")
+            if [ ! -z "${ALT_TEXT}" ];then
+                Limgurl=$(echo "--media ${Outfile} --description \'${ALT_TEXT}\'")
+            else
+                Limgurl=$(echo "--media ${Outfile} --description \'An automated image pulled from the post - ${title}\'")
+            fi            
         else
             Limgurl=""
         fi
